Minor changes in PR

+++ rocAL/rocAL/include/pipeline/image.h
@@ -105,7 +105,7 @@ struct ImageInfo
     Type type() const { return _type; }
     unsigned batch_size() const {return _batch_size;}
     RocalMemType mem_type() const { return _mem_type; }
-    unsigned data_size() const { return _data_size; }
+    uint64_t data_size() const { return _data_size; }
     RocalColorFormat color_format() const {return _color_fmt; }
     unsigned get_roi_width(int image_batch_idx) const;
     unsigned get_roi_height(int image_batch_idx) const;
@@ -119,7 +119,7 @@ private:
     unsigned _height;//!< image height for a single image in the batch
     unsigned _color_planes;//!< number of color planes
     unsigned _batch_size;//!< the batch size (images in the batch are stacked on top of each other)
-    unsigned _data_size;//!< total size of the memory needed to keep the image's data in bytes including all planes
+    uint64_t _data_size;//!< total size of the memory needed to keep the image's data in bytes including all planes
     RocalMemType _mem_type;//!< memory type, currently either OpenCL or Host
     RocalColorFormat _color_fmt;//!< color format of the image
     std::shared_ptr<std::vector<uint32_t>> _roi_width;//!< The actual image width stored in the buffer, it's always smaller than _width/_batch_size. It's created as a vector of pointers to integers, so that if it's passed from one image to another and get updated by one and observed for all.
